rename route, component, password requirements
